Remove potential retain cycle from timer
I'm not sure this is needed. I found one article that says a retain cycle
is definitely created from this.

However when I introduce a `deinit` in `SleepViewController`, the VC
deinitialises fine. Makes me think that a retain cycle isn't being created
here. But just in case it is, I'll add the `[weak self]`